requirements.txt
fastapi
google-auth
google-auth-oauthlib
matplotlib
numpy
openai
pandas
pydantic
python-dotenv
requests
sqlalchemy
uvicorn
